backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers
pillow
django-filter
numpy
djangorestframework-simplejwt>=5.2.0,<6.0
reportlab